requirements.txt
streamlit
supabase
Pillow
